yomo_backend/requirements.txt
fastapi
uvicorn
cohere
requests
python-dotenv
